feat!: reorganize CLI commands
"info" is no more - it's been split up and moved.
- Default "info" output has been moved to the default "chart-review"
  mode without any subcommand.
- "info --ids" has been moved to a toplevel "ids" subcommand.
- "info --labels" has been moved to a toplevel "labels" subcommand.
