+++ chart_review/__init__.py
@@ -1,3 +1,3 @@
 """Chart Review public entry point"""
 
-__version__ = "1.3.0"
+__version__ = "2.0.0"
